+++ compose/network.py
@@ -15,7 +15,8 @@ log = logging.getLogger(__name__)
 
 class Network(object):
     def __init__(self, client, project, name, driver=None, driver_opts=None,
-                 ipam=None, external_name=None):
+                 ipam=None, external_name=None, internal=False, enable_ipv6=False,
+                 labels=None):
         self.client = client
         self.project = project
         self.name = name
@@ -23,6 +24,9 @@ class Network(object):
         self.driver_opts = driver_opts
         self.ipam = create_ipam_config_from_dict(ipam)
         self.external_name = external_name
+        self.internal = internal
+        self.enable_ipv6 = enable_ipv6
+        self.labels = labels
 
     def ensure(self):
         if self.external_name:
@@ -68,6 +72,9 @@ class Network(object):
                 driver=self.driver,
                 options=self.driver_opts,
                 ipam=self.ipam,
+                internal=self.internal,
+                enable_ipv6=self.enable_ipv6,
+                labels=self.labels,
             )
 
     def remove(self):
@@ -115,6 +122,9 @@ def build_networks(name, config_data, client):
             driver_opts=data.get('driver_opts'),
             ipam=data.get('ipam'),
             external_name=data.get('external_name'),
+            internal=data.get('internal'),
+            enable_ipv6=data.get('enable_ipv6'),
+            labels=data.get('labels'),
         )
         for network_name, data in network_config.items()
     }