+++ script/run/run.sh
@@ -15,7 +15,7 @@
 
 set -e
 
-VERSION="1.8.1"
+VERSION="1.9.0-rc1"
 IMAGE="docker/compose:$VERSION"